requirements.txt
streamlit_tags
flask
airtable
pyairtable
openai==1.61.0
urllib3
streamlit
pycountry
pandas
numpy
matplotlib
seaborn
reportlab
PyPDF2
apify-client
langchain_community
langchain_openai
langchain_core
langchain
langchain-chroma
transformers
pinecone
langchain_pinecone
airtable-python-wrapper
unstructured
tiktoken
pypdf
python-dotenv
google_auth_oauthlib
google-auth-oauthlib
requests
Flask-SQLAlchemy
psycopg2-binary
sshtunnel
anthropic
torch
python-docx
pdfplumber
openpyxl
pytesseract 
pillow
notion_client